omniatlas/requirements.txt
scrapy==2.13.2
itemloaders==1.3.2
mysql-connector-python==9.3.0
